fix: build

+++ src/models/Product.model.ts
@@ -11,7 +11,7 @@ class Product extends Model {
     declare name: string
 
     @Column({
-        type: DataType.FLOAT(6, 2)
+        type: DataType.DECIMAL(6, 2)
     })
     declare price: number
